requirements.txt
flask
transformers
torch
textblob
librosa
openai-whisper
